requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3

